Move LogQuery to separate file to lower mem usage
